requirements.txt
pydantic
yt-dlp
# openai
langchain
langchain-openai
pandas
opencv-python
scenedetect
pyyaml
numpy
ffmpeg-python
tqdm
moviepy